refactor: updatcms paragraph
