+++ src/app/components/Cms/CmsParagraph/CmsParagraphEditor.tsx
@@ -3,7 +3,7 @@ import styles from './CmsParagraphEditor.module.scss'
 import EditOverlay from '@/components/Cms/EditOverlay'
 import { EditModeContext } from '@/context/EditMode'
 import Form from '@/components/Form/Form'
-import { updateCmsParagraph } from '@/actions/cms/paragraphs/update'
+import { updateCmsParagraphAction } from '@/actions/cms/paragraphs/update'
 import PopUp from '@/components/PopUp/PopUp'
 import { useContext, useState } from 'react'
 import { useRouter } from 'next/navigation'
@@ -47,7 +47,7 @@ export default function CmsParagraphEditor({ cmsParagraph, editorClassName }: Pr
             <div className={`${styles.CmsParagraphEditor} ${editorClassName}`}>
                 <DynamicSimpleMDEditor className={styles.editor} value={content} onChange={handleContentChange} />
                 <Form
-                    action={updateCmsParagraph.bind(null, cmsParagraph.id).bind(null, content)}
+                    action={updateCmsParagraphAction.bind(null, cmsParagraph.id).bind(null, content)}
                     submitText="Oppdater"
                     successCallback={() => {
                         refresh()
